remove qikprop from default setup

+++ cgi-bin/computations.py
@@ -32,13 +32,14 @@ q.close()
 dbconn.close()
 
 ##### RUN QIKPROP##############
-os.chdir(cgidir+'/qikprop')
-shutil.copyfile('../../public/uploads/structures/'+str(molid)+'-3d.mol',os.getcwd()+'/'+str(molid)+'-3d.mol')
-subprocess.call(['./qikprop',str(molid)+'-3d.mol'], stdout=open(os.devnull,'w'),stderr=open(os.devnull,'w'))
-shutil.copyfile('QP.out','../../public/uploads/qikprop/'+str(molid)+'-QP.txt')
-for i in ['QP.out','QPmyfits','QPwarning','Similar.name','QP.CSV','QPSA.out',str(molid)+'-3d.mol']:
-    try:
-        os.remove(i)
-    except Exception:
-        pass
+if(os.path.isdir('qikprop')):
+    os.chdir(cgidir+'/qikprop')
+    shutil.copyfile('../../public/uploads/structures/'+str(molid)+'-3d.mol',os.getcwd()+'/'+str(molid)+'-3d.mol')
+    subprocess.call(['./qikprop',str(molid)+'-3d.mol'], stdout=open(os.devnull,'w'),stderr=open(os.devnull,'w'))
+    shutil.copyfile('QP.out','../../public/uploads/qikprop/'+str(molid)+'-QP.txt')
+    for i in ['QP.out','QPmyfits','QPwarning','Similar.name','QP.CSV','QPSA.out',str(molid)+'-3d.mol']:
+        try:
+            os.remove(i)
+        except Exception:
+            pass
